Rename some files and functions
