+++ frontend/components/GamePage.tsx
@@ -20,8 +20,8 @@ import { useAuthContext } from '@/contexts/AuthContext'
 import { deserializeToInstance, deserializeToInstances } from '@/functions/serialize'
 import Collection from '@/classes/Collection'
 import styles from "./GamePage.module.css"
-import SenatorsTab from '@/components/MainTab_Senators'
-import FactionsTab from '@/components/MainTab_FactionsTab'
+import SenatorsTab from '@/components/MainSection_SenatorsTab'
+import FactionsTab from '@/components/MainSection_FactionsTab'
 import DetailSection from '@/components/DetailSection'
 import Turn from '@/classes/Turn'
 import Phase from '@/classes/Phase'
